requirements.txt
absl-py==2.1.0
alembic==1.14.1
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dacite==1.9.2
dm-tree==0.1.8
docker==7.1.0
entrypoints==0.4
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.70.0
h5py==3.13.0
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
importlib_metadata==7.2.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras==2.15.0
keras-nlp==0.6.1
keras-core==0.1.7
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mido==1.3.3
ml-dtypes==0.2.0
mlflow==2.11.1
multimethod==1.12
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==23.2
pandas==2.2.0
pandas-profiling
patsy==1.0.1
phik==0.12.4
pillow==11.1.0
pretty-midi==0.2.10
protobuf==4.25.6
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.6.0
#pywin32==308 #Windows
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.38
sqlparse==0.5.3
statsmodels==0.14.4
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
visions==0.7.6
waitress==3.0.2
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.14.1
ydata-profiling==4.12.2
zipp==3.21.0
